+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+**v1.1.0**
+- upated sns-payload-validator to v1.1.0 to support SignatureVersion 2
+- add SignatureVersion 2 tests
+- Updated README.md
+
 **v1.0.1**
 - fix reposizitory in package.json
 - better wording in README.md